+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Optionally allow web lookup of title names via the new `-AllowWebLookup` parameter (#3).
 - Provide helpful user suggestions when title names cannot be resolved (#4).
+- Exit early if not on Windows (#5).
 
 ### Changed